lint: pay down the simply-elegant ruff suite to zero
ruff (select = ALL) over src/tests/scripts under the plugin's config went
from 10,285 findings to 0. pyproject.toml carries every project-level
overrule with its reason beside it, merged onto the plugin config by the
hooks; the rest is fixed in code: safe and selected unsafe autofixes, then
hand fixes and codemods (append-loop -> extend, Union aliases -> |, long
string literals split, pytest match= patterns raw or escaped).

Two accidental contracts the autofixes broke and this commit restores:
healing_helpers re-exported three slotlib functions that 33 champion modules
read through the module (callers now import slotlib directly), and a test's
unused import was the front door D-95 counts (kept, with its noqa reason).
Two real bugs surfaced: a test whose implicit string concatenation split an
assert into an always-true literal and a bare comparison, and a forward
reference an unnecessary-lambda fix would have turned into a NameError.

Prose: the history scan's 17 non-HANDOVER hits reworded to present state.
Gates: pytest 15370 passed, golden snapshot identical, black clean.

+++ src/calculator/champions/aurelion_sol.py
@@ -61,6 +61,8 @@ from typing import Any
 from ..ability_spec import DamagePart
 from ..binary_roots import data_value, spell_object
 from .engine import CC_PER_PART, SlotCtx, build_parser
+from .inputs import bool_option, int_option
+from .module_contract import coverage
 from .module_helpers import delayed_damage, no_damage
 from .slotlib import (
     ability_name,
@@ -71,8 +73,6 @@ from .slotlib import (
     extract_value,
 )
 from .source_receipts import load_champion_sources
-from .inputs import bool_option, int_option
-from .module_contract import coverage
 
 # One full Q channel: 3.25 s of beam, with a burst on the primary target
 # at each full second of channel (3 bursts).
@@ -260,7 +260,7 @@ def _breath_of_light(ctx: SlotCtx) -> dict[str, Any] | None:
     # The beam is per-tick damage x (seconds / tick interval): 26 ticks
     # of the per-tick row for one 3.25s channel, exactly the sourced
     # "Total Maximum Magic Damage" (per-second x 3.25).
-    ticks = int(round(seconds / _Q_TICK_INTERVAL))
+    ticks = round(seconds / _Q_TICK_INTERVAL)
     per_tick = beam_per_second * _Q_TICK_INTERVAL
     total = per_tick * ticks + per_burst * bursts
     parts = [
